Bullets, lives, waves and misc fixes
When a bullet hits an enemy or a player, they lose a life. If an enemy's life drops below 0, it is removed from the Vector of enemies in the scene, if the player's lives drop below 0, the game is ended. Static enemies fire bullets to the bottom of the screen at random. Implemented waves that increase in difficulty through number of enemies, number of lives those enemies have and the chance that they fire a bullet in a given update loop. Fixed bug where the top bar would return if you minimised and then reopened the game. Added delta time constraints to the game loop, much more responsive. Finally, showed the lives and wave number on screen and commented and cleaned up all of the code.

+++ app/src/main/java/Model/DrawableObject.java
@@ -11,19 +11,19 @@ import android.graphics.drawable.Drawable;
 * a resources from the drawables folder.*/
 public class DrawableObject
 {
-    protected Vector2i m_pos; //The left and top of this object's bounding rectangle.
-    private Vector2i m_scale; //The right and bottom of this object's bounding rectangle.
+    protected Vector2f m_pos; //The left and top of this object's bounding rectangle.
+    private Vector2f m_scale; //The right and bottom of this object's bounding rectangle.
 
     private Drawable m_sprite; //Reference to the image passed when this object is constructed.
 
-    public DrawableObject (Drawable image, Vector2i position, Vector2i scale)
+    public DrawableObject (Drawable image, Vector2f position, Vector2f scale)
     {
         m_pos = position;
         m_scale = scale;
         m_sprite = image;
 
         /*Setup the bounding rectangle of this object.*/
-        m_sprite.setBounds(m_pos.x, m_pos.y, m_pos.x + m_scale.x, m_pos.y + m_scale.y);
+        m_sprite.setBounds((int)m_pos.x, (int)m_pos.y, (int)(m_pos.x + m_scale.x), (int)(m_pos.y + m_scale.y));
     }
 
     /*Change the Drawable that this object's sprite references.*/
@@ -32,23 +32,23 @@ public class DrawableObject
     }
 
     /*Set this object's position to its current positon plus x and y.*/
-    public void translate(int x, int y)
+    public void translate(float x, float y)
     {
-        setPosition(new Vector2i(m_pos.x + x, m_pos.y + y));
+        setPosition(new Vector2f(m_pos.x + x, m_pos.y + y));
     }
 
     /*Set this object's position outright to the given Vector2.*/
-    public void setPosition(Vector2i position)
+    public void setPosition(Vector2f position)
     {
         m_pos = position;
-        m_sprite.setBounds(m_pos.x, m_pos.y, m_pos.x + m_scale.x, m_pos.y + m_scale.y);
+        m_sprite.setBounds((int)m_pos.x, (int)m_pos.y, (int)(m_pos.x + m_scale.x), (int)(m_pos.y + m_scale.y));
     }
 
     /*Set the width and height of this objects bounding rectangle.*/
-    public void setScale(Vector2i scale)
+    public void setScale(Vector2f scale)
     {
         m_scale = scale;
-        m_sprite.setBounds(m_pos.x, m_pos.y, m_pos.x + m_scale.x, m_pos.y + m_scale.y);
+        m_sprite.setBounds((int)m_pos.x, (int)m_pos.y, (int)(m_pos.x + m_scale.x), (int)(m_pos.y + m_scale.y));
     }
 
     /*Draw this object's sprite on the given canvas.*/
@@ -62,7 +62,7 @@ public class DrawableObject
         return m_sprite.getBounds();
     }
 
-    public Vector2i getPos()
+    public Vector2f getPos()
     {
         return m_pos;
     }
